requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.5
chardet==3.0.4
discord.py==1.6.0
idna==3.1
multidict==5.1.0
opuslib==3.0.1
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.15.0
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
youtube_dl==2021.6.6
gtts==2.2.3